requirements.txt
openai
composio-openai
google-cloud 
google-auth 
google-auth-oauthlib